feat: increased time variable precision to long double feat: added verbose command to make, changed all command to be less verbose

+++ include/parallel.h
@@ -3,10 +3,10 @@
 
 #include <stdbool.h>
 
-bool is_symmetric_omp(float **matrix, int n, double* time);
+bool is_symmetric_omp(float **matrix, int n, long double* time);
 
-float** transpose_omp(float **matrix, int n, double* time);
+float** transpose_omp(float **matrix, int n, long double* time);
 
-float** transpose_omp_block_based(float **matrix, int n, int block_size, double* time);
+float** transpose_omp_block_based(float **matrix, int n, int block_size, long double* time);
 
 #endif // !PARALLEL_H
